Move gh_token_encryption code to crates_io_encryption crate
Done via `git mv`
