+++ src/util.rs
@@ -1,10 +1,10 @@
 pub use self::io_util::{read_fill, read_le_u32};
 pub use self::request_helpers::*;
 pub use crates_io_database::utils::token;
+pub use crates_io_encryption as gh_token_encryption;
 
 pub mod diesel;
 pub mod errors;
-pub mod gh_token_encryption;
 mod io_util;
 pub mod oauth;
 mod request_helpers;
